Added castling functionality
The functionality for castling is executed by initially verifying whether the king has already made a move. Subsequently, the player's turn is examined, followed by an assessment of whether the rooks are positioned as required. If the rooks are in their designated starting positions and have not yet moved, the castling maneuver is performed accordingly.